+++ Consoles/LayeredConsole.cs
@@ -246,31 +246,14 @@ namespace SadConsoleEditor.Consoles
                 _layerMetadata[i].Index = i;
         }
 
-        public static void Save(LayeredConsole instance, string file)
+        public void Save(string file)
         {
-            if (System.IO.File.Exists(file))
-                System.IO.File.Delete(file);
-
-            var serializer = new System.Runtime.Serialization.Json.DataContractJsonSerializer(typeof(LayeredConsole));
-            using (var stream = System.IO.File.OpenWrite(file))
-            {
-                serializer.WriteObject(stream, instance);
-            }
+            SadConsole.Serializer.Save<LayeredConsole>(this, file, new Type[] { typeof(Console) });
         }
 
         public static LayeredConsole Load(string file)
         {
-            if (System.IO.File.Exists(file))
-            {
-                using (var fileObject = System.IO.File.OpenRead(file))
-                {
-                    var serializer = new System.Runtime.Serialization.Json.DataContractJsonSerializer(typeof(LayeredConsole));
-
-                    return serializer.ReadObject(fileObject) as LayeredConsole;
-                }
-            }
-
-            throw new System.IO.FileNotFoundException("File not found.", file);
+            return SadConsole.Serializer.Load<LayeredConsole>(file);
         }
     }